// Bank-gui/src/app/model/customer.model.ts
export class Customer {
    // customerid: any;
    firstname: string;
    middlename: string;
    lastname: string;
    dateofbirth: string;
    gender: string;
    email: string;
    password: string;
    contact: string;
    pan: string;
    nationality: string;
    salary: any;
    street: string;
    city: string;
    state: string;
    pin: string;
}